requirements.txt
click
PyGithub
python-dotenv
requests